Trying level loader class
We can use a hashmap to store level data and use this level loader class
to... well load the levels as their selected.

Once a level has been selected, we search the hash for the level,
LevelLoader loads it up with the necessary Box2D fixtures. This
implementation will be TMX files for now.